+++ NEWS.md
@@ -1,3 +1,5 @@
+# DO.utils (development version)
+
 # DO.utils 0.3.0
 
 ## General